+++ bauh/api/abstract/handler.py
@@ -17,9 +17,10 @@ class ProcessWatcher:
         """
         pass
 
-    def request_confirmation(self, title: str, body: Optional[str], components: List[ViewComponent] = None, confirmation_label: str = None,
-                             deny_label: str = None, deny_button: bool = True, window_cancel: bool = False,
-                             confirmation_button: bool = True) -> bool:
+    def request_confirmation(self, title: str, body: Optional[str], components: List[ViewComponent] = None,
+                             confirmation_label: str = None, deny_label: str = None, deny_button: bool = True,
+                             window_cancel: bool = False, confirmation_button: bool = True,
+                             min_width: Optional[int] = None) -> bool:
         """
         request a user confirmation. In the current GUI implementation, it shows a popup to the user.
         :param title: popup title
@@ -30,6 +31,7 @@ class ProcessWatcher:
         :param deny_button: if the deny button should be displayed
         :param window_cancel: if the window cancel button should be visible
         :param confirmation_button: if the confirmation button should be displayed
+        :param min_width: minimum width for the confirmation dialog
         :return: if the request was confirmed by the user
         """
         pass