+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.X.X]
+
+### Improvements
+- UI
+  - upgrade summary dialog size
+
 ## [0.9.28] 2022-02-14
 
 ### Fixes